Fixed Eclipse plugin install issues

+++ Plugins/Aspose_Pdf_Java_Maven_for_Eclipse/AsposePdfEclipsePlugin/src/com/aspose/pdf/maven/examples/AsposeExampleSupport.java
@@ -34,9 +34,9 @@ public class AsposeExampleSupport implements IRunnableWithProgress {
 	private String exampleCategory;	
 	private IProject project;
 
-	final static String localExampleFolder = "aspose/GitConsRepos/Aspose.Pdf/Examples";
+	final static String localExampleFolder = "aspose/GitConsRepos/Aspose.PDF/Examples";
 	final static String localExampleSourceFolder = "src/main/java/com/aspose/pdf/examples";
-	final static String localExampleResourceFolder = "src/main/resources/com/aspose/pdf/examples";
+	final static String localExampleResourceFolder = "src/main/resources/com/aspose/pdf/examples";	
 
 	public AsposeExampleSupport(String selectedProjectPath, String exampleCategory, IProject project) {
 		this.selectedProjectPath = selectedProjectPath;
@@ -74,9 +74,15 @@ public class AsposeExampleSupport implements IRunnableWithProgress {
 				+ File.separator + AsposeConstants.MAVEN_POM_XML;
 
 		try {						
-			FileUtils.copyDirectory(srcExampleCategoryPath, destExampleCategoryPath);
-			Files.copy(srcUtil, destUtil, StandardCopyOption.REPLACE_EXISTING);
-			FileUtils.copyDirectory(srcExampleResourceCategoryPath, destExampleResourceCategoryPath);
+			if(srcExampleCategoryPath.exists()) {
+				FileUtils.copyDirectory(srcExampleCategoryPath, destExampleCategoryPath);	
+			}					
+			if(srcUtil.toFile().exists()) {
+				Files.copy(srcUtil, destUtil, StandardCopyOption.REPLACE_EXISTING);	
+			}			
+			if(srcExampleResourceCategoryPath.exists()) {
+				FileUtils.copyDirectory(srcExampleResourceCategoryPath, destExampleResourceCategoryPath);				
+			}
 
 			NodeList examplesNoneAsposeDependencies = AsposeMavenProjectManager.getInstance()
 					.getDependenciesFromPOM(repositoryPOM_XML, AsposeConstants.ASPOSE_GROUP_ID);
